+++ packages/react/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @writeme-ui/react
 
+## 1.4.0
+
+### Minor Changes
+
+- Added Semantic Colors
+  Added Button Variation Danger
+
+## 1.3.1
+
+### Patch Changes
+
+- Add option to pass reference to TextInput and adjust some styles
+
 ## 1.3.0
 
 ### Minor Changes