+++ packages/tokens/CHANGELOG.md
@@ -4,4 +4,11 @@
 
 ### Minor Changes
 
+- Added Semantic Colors
+  Added Button Variation Danger
+
+## 1.2.0
+
+### Minor Changes
+
 - a2b8914: Added core components